kcov: Use RT*alpha when calculating binding strength.
In the continuing saga of alpha being confusing, I made a mistake in how
I thought it should be handled for sequence dependence.  Here, we need
alpha to be unitful, and in Gibbs energy (kcal/mol) units, if we have
a unitful Gibbs energy instead of kTAM's unitless one.

Longer term, it would probably be better to (a) have me insist to everyone
that the unitless value I have been pushing to everyone for the past
several years should be unitful and in entropic units (eg, kcal/mol/K),
(b) have us rename it when we're using unitful energies, or (c) have
us handle the entropic penalty differently.